+++ packages/unlock-js/src/index.ts
@@ -12,4 +12,3 @@ export * from './subgraph'
 export * from './erc20'
 export * from './erc20abi'
 export * from './KeyManager'
-export * from './customHookService'
